+++ src/include/setup.hpp
@@ -41,12 +41,17 @@ class Setup {
 
   static int64_t getUpdateIntervalInDays();
 
+  static std::string getLatestVersion();
+
+  static void checkLatestedVersion();
+
   static bool hasMemory();
 
   static void runWizard();
 
  private:
-  static std::string USER_HOME;
+  static constexpr int64_t DEFAULT_UPDATE_INTERVAL = 1;  // 1 day default
+
   static std::string LODA_HOME;
   static std::string OEIS_HOME;
   static std::string PROGRAMS_HOME;
@@ -71,6 +76,7 @@ class Setup {
   static bool checkMineParallelScript();
   static bool checkSubmittedBy();
   static bool checkMaxMemory();
+  static bool checkUpdateInterval();
 
   static bool updateFile(const std::string& local_file, const std::string& url,
                          const std::string& header, const std::string& marker,